FIRST SUCCESSFUL CONNECTION TO SERVER

+++ app/src/main/java/com/example/asus/example/mvvm/Model/Entities/User.java
@@ -11,7 +11,7 @@ import java.util.List;
  */
 public class User implements Serializable {
 
-    private long id;
+    private int id;
     private String name;
     private String email;
     private String imageUrl;
@@ -54,7 +54,7 @@ public class User implements Serializable {
      * @param createdEvents all the Events, the User created.
      * @param createdComments all the Comments, the User created.
      */
-    public User(long id, String name, String email, String imageUrl, List<User> subscriptions, List<User> subscribers, List<Group> joinedGroups, List<Event> participatedEvents, List<Post> createdPosts, List<Group> createdGroups, List<Event> createdEvents, List<Comment> createdComments) {
+    public User(int id, String name, String email, String imageUrl, List<User> subscriptions, List<User> subscribers, List<Group> joinedGroups, List<Event> participatedEvents, List<Post> createdPosts, List<Group> createdGroups, List<Event> createdEvents, List<Comment> createdComments) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -104,7 +104,7 @@ public class User implements Serializable {
      * method to get the id of the User
      * @return id of the User
      */
-    public long getId() {
+    public int getId() {
         return id;
     }
 
@@ -112,7 +112,7 @@ public class User implements Serializable {
      * sets the id of the User
      * @param id if of the User
      */
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
